bridge/requirements.txt
fastapi==0.94.0
uvicorn==0.21.0
pydantic==1.10.6
alembic==1.10.2
psycopg2==2.9.5
asyncpg==0.27.0
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
fastapi-users[sqlalchemy]
sqlalchemy
